Next Permutation (LeetCode - 31)
You are given an array of integers representing a permutation of numbers. The goal is to rearrange the array into the next lexicographically greater permutation. If such a permutation is not possible (the array is sorted in descending order), the array should be rearranged to the smallest possible permutation (sorted in ascending order). The modification must be done in place, with constant extra memory.

Example:  
Input: nums = [1,2,3]  
Output: [1,3,2]  

Input: nums = [3,2,1]  
Output: [1,2,3]  

Input: nums = [1,1,5]  
Output: [1,5,1]  

Constraints:  
- 1 <= nums.length <= 100  
- 0 <= nums[i] <= 100  

The task is to rearrange the elements of the array to the next permutation in-place, or to the smallest permutation if no next permutation exists.