Merge Sorted Array (LeetCode - 88)
Merge Sorted Array

You are given two integer arrays nums1 and nums2, sorted in non-decreasing order, and two integers m and n, representing the number of elements in nums1 and nums2, respectively.

nums1 has a length of m + n, where the first m elements represent the elements to merge, and the last n elements are set to 0 and should be ignored. nums2 has a length of n.

Write a function that merges nums2 into nums1 as one sorted array.

The final sorted array should not be returned by the function but should be stored inside the array nums1. To accommodate this, nums1 has enough space (size m + n) to hold the elements of both arrays.

Example:

Input:
nums1 = [1,2,3,0,0,0], m = 3
nums2 = [2,5,6], n = 3

Output:
[1,2,2,3,5,6]

Explanation:
The elements of nums2 are merged into nums1 in sorted order.

Constraints:
- nums1.length == m + n
- nums2.length == n
- 0 <= m, n <= 200
- -10^9 <= nums1[i], nums2[j] <= 10^9

Goal: Implement the function to merge the arrays in place.