Count Inversions
Count Inversions in an Array

You are given an array of integers. An inversion is a pair of indices (i, j) such that i < j and arr[i] > arr[j]. Your task is to count the total number of inversions in the array.

Example:

Input:
arr = [5, 3, 2, 4, 1]

Output:
8

Explanation:
The inversions are: (0,1), (0,2), (0,3), (0,4), (1,2), (1,4), (2,4), (3,4)

Constraints:
- 1 <= n <= 10^5
- -10^9 <= arr[i] <= 10^9

Goal: Implement an efficient algorithm to count the inversions.