Search A 2D Matrix (LeetCode - 74)
Search a 2D Matrix

You are given an m x n integer matrix sorted in ascending order, both row-wise and column-wise. Your task is to write a function that determines if a given target value exists in the matrix.

The matrix has the following properties:
- Integers in each row are sorted from left to right.
- The first integer of each row is greater than the last integer of the previous row.

Example:

Input:
matrix = [
  [1, 3, 5, 7],
  [10, 11, 16, 20],
  [23, 30, 34, 60]
]
target = 3

Output:
true

Explanation:
The number 3 exists in the matrix.

Constraints:
- m == matrix.length
- n == matrix[i].length
- 1 <= m, n <= 100
- -10^4 <= matrix[i][j], target <= 10^4

Goal: Implement an efficient algorithm to search for the target value in the matrix.