Delete Node in a Linked List (LeetCode - 237)
**Delete Node in a Linked List**

You are given a node in a singly linked list. This node is not the tail and is guaranteed to be part of a valid linked list. Your task is to delete this node from the list, but you do not have access to the head of the list.

Since you can’t access the node before the one to be deleted, you need to delete it by copying the value of the next node into the current node and then bypassing the next node.

**Example:**

Input: Linked list = 4 → 5 → 1 → 9  
Node to delete = 5  
Output: 4 → 1 → 9

**Constraints:**

- The number of nodes in the list is between 2 and 1000.  
- The value of each node is between 0 and 10000.  
- The node to be deleted is not the tail node.  
- All node values are unique.