Reverse Nodes in K Group (LeetCode - 25)
**Problem: Reverse Nodes in k-Group**

Given the head of a singly linked list and an integer k, reverse the nodes of the list k at a time and return the modified list. If the number of nodes is not a multiple of k, leave the last remaining nodes as they are. You are not allowed to change the values of the nodes, only the nodes themselves can be changed.

**Example:**

Input: head = [1, 2, 3, 4, 5], k = 2  
Output: [2, 1, 4, 3, 5]

**Constraints:**

- The number of nodes in the list is n  
- 1 ≤ n ≤ 5000  
- 0 ≤ Node value ≤ 1000  
- 1 ≤ k ≤ 5000